requirements.txt
kafka-python>=2.0.2,<3
PyYAML>=6.0
python-snappy>=0.6